requirements.txt
Django~=3.2.6
Pillow~=8.3.2